WithRange: demote.
This thing was accumulating too many functions which were each only used
by one or two of its consumers. Scrub of magic, moving those functions
into the actual places they're relied upon

The reason this was embarked upon was to let Token/Stmt/Expr/etc.
actually be defined explicitly, so they can be changed to file-level
structs if it ends up making sense to do so.