adc: separate reader thread.
Necessary? We could just process the debugs in the reader return. That
said, then we won't get them until we next look for them, but that might
be perfectly acceptable too.

At some point this'll just be part of an event loop and we won't really
need to think about it, so just leave it until it becomes something else
or is easily refactored.